refactoring completed
