+++ .gitignore
@@ -51,3 +51,6 @@ qtcreator-*
 
 # Catkin custom files
 CATKIN_IGNORE
+
+# vscode
+.vscode
