+++ src/amr_gazebo/launch/linked_drive_error_test.launch
@@ -3,9 +3,9 @@
 <launch>
 
     <!-- # Linked Drive -->
-    <node pkg="linked_diff_drive" type="linked_drive_pub2.py" name="linked_drive_node" output="screen"/>
+    <node pkg="linked_diff_drive" type="linked_drive_pub.py" name="linked_drive_node" output="screen"/>
 
     <!-- # Follower Error Plot -->
-    <!-- <node pkg="amr_gazebo" type="linked_drive_error.py" name="error_plot_node" output="screen"/> -->
+    <node pkg="amr_gazebo" type="linked_drive_error.py" name="error_plot_node" output="screen"/>
 
 </launch>
